Fix tests that are failing with Node's v0.12 branch.
Node's v0.12 branch will bind server sockets to '::' by
default if IPv6 is available on the host and if no host
is provided.

Thus, some tests that check for either servers' or clients' IP
addresses and that assume IPv4 only addresses would fail.

This change is backward compatible with Node 0.10.x.