Update README.md
Hi,

I added section "Books" and a link to the hapi.js in Action written by Matt Harrison for Manning Publications. We believe this book is a great resource and all members of community will benefit from this information. If you find this inappropriate, please let me know where I can publish it.