Add a new 'started' event, emitted when start callbacks are complete
Rename test variables started -> start to denote the semantic difference
between the 2 events.

Update the documentation.

This new event is useful to allow users to determine if the server
(including any bespoke initialisation in start's callback) is started.

This change was prompted by the repeated need to implement the following:

server.start(err => {
  // init that could fail.

  server.emit('started');
});

server.on('started', () => {
  log('server startup complete');
});