Invoke reply with promise returned from handler.
This allows a promise to be returned from a handler which will then
automatically be used to invoke `reply`. This way the server responds as
soon as the promise resolves or rejects.

This cleans up the code in handlers that uses promises significantly.

Simple example:
```
function handle (request, reply) {

    const promise = Promise.resolve(request.params.id);

    reply(promise);
};
```

Becomes:

```
function handle (request, reply) {

    return Promise.resolve(request.params.id);
};
```