Allow undefined handler and prehandler response values
There isn't anything inherently wrong with a handler or pre-handler that
doesn't need to modify the response. As an example, a pre-handler may
simply enforce RBAC privileges:

const checkWriteAccess = request => {
  if (request.params.admin) {
    throw Boom.unauthorized();
  }
}

(This is admittedly bad security design, but imagine a proper RBAC
mechanism)

Likewise, a handler that doesn't return a value may be used for a route
that returns an empty 204 response.