fix: 🐛 typings around server.decorate (#4538)
* Update decorator types to be correct

- The code currently allows any value to be added as a decorator, but the
types do not. This work allows any value to be added to a decorator by
updating the types to reflect what the code does
- Add relevant tests
- Update inline docs
- Create and export DecorationValue type
- Fixes #4524

* Update work to add extra decorate definitions

These extra definitions specify how to use value instead of method

* fix: 🐛 typings around server.decorate

Types match the validation a bit more. Type errors with error when
attempting to decorate a builtin.

---------

Co-authored-by: Ben Chidgey <ben@feedmypixel.com>